feat: two-tier server architecture for CLI auth
- Add /.well-known/ephemeral-auth discovery endpoint
- Support explicit auth_method (token/web) in response
- Platform servers can redirect CLI to separate git server
- Remove WebAuthURL from server config (platform handles this)

+++ internal/store/sqlite_test.go
@@ -205,7 +205,7 @@ func TestStore_RepoLifecycle(t *testing.T) {
 		require.NoError(t, s.UpdateRepoSize("repo-1", 2048))
 
 		got, _ := s.GetRepoByID("repo-1")
-		assert.Equal(t, 2048, got.SizeBytes)
+		assert.Equal(t, int64(2048), got.SizeBytes)
 	})
 
 	t.Run("list", func(t *testing.T) {
